chore: moved `Icon` component to the `uikit` directory; added `eye` icons
