+++ src/components/common/index.ts
@@ -1,3 +1,2 @@
 export { GoBackButton } from "./GoBackButton";
 export { TagClosable } from "./TagClosable";
-export { Icon } from "./icon";
