+++ src/components/uikit/icon/index.ts
@@ -1,0 +1,2 @@
+export { Icon } from "./icon";
+export type { SpritesMap } from "./sprite.gen";
