+++ src/components/uikit/index.ts
@@ -20,3 +20,4 @@ export * from "./Radio";
 export * from "./Tabs";
 export * from "./filters";
 export * from "./Forms";
+export * from "./icon";
